+++ src/main/java/bridge/controller/Controller.java
@@ -1,11 +1,10 @@
 package bridge.controller;
 
-import bridge.OutputView;
+import bridge.view.OutputView;
 import bridge.domain.Bridge;
 import bridge.domain.BridgeGame;
 import bridge.domain.BridgeMaker;
 import bridge.domain.GameStates;
-import bridge.dtos.GameResultDto;
 import bridge.dtos.MapStateDto;
 import bridge.enums.BridgeCell;
 import bridge.enums.GameCommand;
