+++ src/main/java/bridge/view/InputView.java
@@ -1,7 +1,6 @@
-package bridge;
+package bridge.view;
 
 import camp.nextstep.edu.missionutils.Console;
-import com.sun.security.jgss.GSSUtil;
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
